Got rid of CashHolderLike!
